* Refactored database code so it uses the GoMySQL package * Changed all integers to "int"

+++ trunk/Server/connection.go
@@ -113,7 +113,7 @@ func (c *Connection) SendPlayerData() {
 	c.SendMessage(readyMessage)
 }
 
-func (c *Connection) SendMapData(_direction uint16, _centerPosition pos.Position) {
+func (c *Connection) SendMapData(_direction int, _centerPosition pos.Position) {
 	xMin := 1
 	xMax := CLIENT_VIEWPORT.X
 	yMin := 1
